requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
db-dtypes==1.4.2
gcloud==0.18.3
google-api-core==2.24.2
google-auth==2.39.0
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
httplib2==0.22.0
idna==3.10
numpy==2.2.4
oauth2client==4.1.3
oauthlib==3.2.2
packaging==25.0
pandas==2.2.3
pandas-gbq==0.28.0
proto-plus==1.26.1
protobuf==6.31.0rc2
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydata-google-auth==1.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
